refactor: extracted common partition and goal functions
