+++ src/chart_types/heatmap/state/selectors/scenegraph.ts
@@ -17,10 +17,10 @@
  * under the License.
  */
 
+import { measureText } from '../../../../common/non_cartesian_utils';
 import { SettingsSpec } from '../../../../specs';
 import { RecursivePartial, mergePartial } from '../../../../utils/common';
 import { Dimensions } from '../../../../utils/dimensions';
-import { measureText } from '../../../partition_chart/layout/utils/measure';
 import { config as defaultConfig } from '../../layout/config/config';
 import { Config } from '../../layout/types/config_types';
 import { ShapeViewModel, nullShapeViewModel } from '../../layout/types/viewmodel_types';
