+++ src/chart_types/partition_chart/layout/utils/treemap.ts
@@ -17,9 +17,9 @@
  * under the License.
  */
 
-import { Pixels } from '../types/geometry_types';
+import { GOLDEN_RATIO } from '../../../../common/constants';
+import { Pixels } from '../../../../common/geometry_types';
 import { Part } from '../types/types';
-import { GOLDEN_RATIO } from './constants';
 import { ArrayEntry, CHILDREN_KEY, entryValue, HierarchyOfArrays } from './group_by_rollup';
 
 const MAX_U_PADDING_RATIO = 0.0256197; // this limits area distortion to <10% (which occurs due to pixel padding) with very small rectangles
