+++ src/chart_types/partition_chart/layout/viewmodel/link_text_layout.ts
@@ -17,14 +17,15 @@
  * under the License.
  */
 
+import { TAU } from '../../../../common/constants';
+import { Distance, PointTuple, PointTuples } from '../../../../common/geometry_types';
+import { TextMeasure } from '../../../../common/non_cartesian_utils';
 import { ValueFormatter, Color } from '../../../../utils/common';
 import { Point } from '../../../../utils/point';
 import { Config } from '../types/config_types';
-import { Distance, PointTuple, PointTuples } from '../types/geometry_types';
-import { Box, Font, TextAlign, TextMeasure } from '../types/types';
+import { Box, Font, TextAlign } from '../types/types';
 import { LinkLabelVM, RawTextGetter, ShapeTreeNode, ValueGetterFunction } from '../types/viewmodel_types';
 import { makeHighContrastColor, integerSnap, monotonicHillClimb, isColorValid } from '../utils/calcs';
-import { TAU } from '../utils/constants';
 import { meanAngle } from '../utils/geometry';
 import { trueBearingToStandardPositionAngle } from '../utils/math';
