+++ src/chart_types/partition_chart/layout/viewmodel/viewmodel.ts
@@ -17,12 +17,14 @@
  * under the License.
  */
 
+import { TAU } from '../../../../common/constants';
+import { Distance, Pixels, PointTuple, Radius } from '../../../../common/geometry_types';
+import { TextMeasure } from '../../../../common/non_cartesian_utils';
 import { StrokeStyle, ValueFormatter, Color } from '../../../../utils/common';
 import { Layer } from '../../specs';
 import { percentValueGetter } from '../config/config';
 import { Config, PartitionLayout } from '../types/config_types';
-import { Distance, Pixels, PointTuple, Radius } from '../types/geometry_types';
-import { TextMeasure, Part, MODEL_KEY } from '../types/types';
+import { Part, MODEL_KEY } from '../types/types';
 import {
   nullShapeViewModel,
   OutsideLinksViewModel,
@@ -35,7 +37,6 @@ import {
   ValueGetterFunction,
 } from '../types/viewmodel_types';
 import { argsToRGBString, stringToRGB } from '../utils/color_library_wrappers';
-import { TAU } from '../utils/constants';
 import { meanAngle } from '../utils/geometry';
 import {
   aggregateAccessor,
