+++ src/chart_types/partition_chart/renderer/canvas/canvas_renderers.ts
@@ -17,9 +17,11 @@
  * under the License.
  */
 
+import { TAU } from '../../../../common/constants';
+import { Pixels } from '../../../../common/geometry_types';
+import { cssFontShorthand } from '../../../../common/non_cartesian_utils';
 import { clearCanvas, renderLayers, withContext } from '../../../../renderers/canvas';
 import { Color } from '../../../../utils/common';
-import { Pixels } from '../../layout/types/geometry_types';
 import {
   LinkLabelVM,
   OutsideLinksViewModel,
@@ -29,8 +31,6 @@ import {
   TextRow,
 } from '../../layout/types/viewmodel_types';
 import { addOpacity } from '../../layout/utils/calcs';
-import { TAU } from '../../layout/utils/constants';
-import { cssFontShorthand } from '../../layout/utils/measure';
 import { LinkLabelsViewModelSpec } from '../../layout/viewmodel/link_text_layout';
 import { isSunburst } from '../../layout/viewmodel/viewmodel';
