+++ src/chart_types/partition_chart/renderer/dom/highlighter.tsx
@@ -19,12 +19,12 @@
 
 import React from 'react';
 
+import { TAU } from '../../../../common/constants';
+import { PointObject } from '../../../../common/geometry_types';
 import { Dimensions } from '../../../../utils/dimensions';
 import { configMetadata } from '../../layout/config/config';
 import { PartitionLayout } from '../../layout/types/config_types';
-import { PointObject } from '../../layout/types/geometry_types';
 import { QuadViewModel } from '../../layout/types/viewmodel_types';
-import { TAU } from '../../layout/utils/constants';
 import { isSunburst, isTreemap } from '../../layout/viewmodel/viewmodel';
 
 /** @internal */
