+++ src/chart_types/partition_chart/specs/index.ts
@@ -20,6 +20,7 @@
 import React from 'react';
 
 import { ChartTypes } from '../..';
+import { Pixels } from '../../../common/geometry_types';
 import { Spec } from '../../../specs';
 import { SpecTypes } from '../../../specs/constants'; // kept as unshortened import on separate line otherwise import circularity emerges
 import { getConnect, specComponentFactory } from '../../../state/spec_factory';
@@ -34,7 +35,6 @@ import {
 } from '../../../utils/common';
 import { config, percentFormatter } from '../layout/config/config';
 import { Config, FillFontSizeRange, FillLabelConfig } from '../layout/types/config_types';
-import { Pixels } from '../layout/types/geometry_types';
 import { ShapeTreeNode, ValueGetter, NodeColorAccessor } from '../layout/types/viewmodel_types';
 import { AGGREGATE_KEY, PrimitiveValue } from '../layout/utils/group_by_rollup';
