+++ src/chart_types/xy_chart/renderer/canvas/primitives/text.ts
@@ -17,10 +17,10 @@
  * under the License.
  */
 
+import { cssFontShorthand, measureText } from '../../../../../common/non_cartesian_utils';
 import { withContext, withRotatedOrigin } from '../../../../../renderers/canvas';
 import { Point } from '../../../../../utils/point';
 import { Font, TextAlign, TextBaseline } from '../../../../partition_chart/layout/types/types';
-import { cssFontShorthand, measureText } from '../../../../partition_chart/layout/utils/measure';
 
 /** @internal */
 export function renderText(
