+++ src/common/constants.ts
@@ -19,3 +19,9 @@
 
 /** @internal */
 export const DEFAULT_CURSOR = 'default';
+/** potential internal */
+export const TAU = 2 * Math.PI;
+/** potential internal */
+export const RIGHT_ANGLE = TAU / 4;
+/** potential internal */
+export const GOLDEN_RATIO = 1.618;
