+++ src/common/geometry_types.ts
@@ -70,4 +70,5 @@ type CirclinePredicateSet = CirclinePredicate[];
 /** @internal */
 export type RingSectorConstruction = CirclinePredicateSet;
 
+/** potential internal */
 export type TimeMs = number;
